Display api: rectangle element

+++ lib/toolbox/value_index.h
@@ -11,7 +11,7 @@ extern "C" {
 /** Get the index of a int32_t array element which is closest to the given value.
  *
  * Returned index corresponds to the first element found.
- * If no suitable elements were found, the function returns 0.
+ * If no suitable elements were found, the function returns -1.
  *
  * @param   value           value to be searched.
  * @param   values          pointer to the array to perform the search in.
@@ -19,12 +19,12 @@ extern "C" {
  *
  * @return value's index.
  */
-size_t value_index_int32(const int32_t value, const int32_t values[], size_t values_count);
+int32_t value_index_int32(const int32_t value, const int32_t values[], size_t values_count);
 
 /** Get the index of a uint32_t array element which is closest to the given value.
  *
  * Returned index corresponds to the first element found.
- * If no suitable elements were found, the function returns 0.
+ * If no suitable elements were found, the function returns -1.
  *
  * @param   value           value to be searched.
  * @param   values          pointer to the array to perform the search in.
@@ -32,12 +32,12 @@ size_t value_index_int32(const int32_t value, const int32_t values[], size_t val
  *
  * @return value's index.
  */
-size_t value_index_uint32(const uint32_t value, const uint32_t values[], size_t values_count);
+int32_t value_index_uint32(const uint32_t value, const uint32_t values[], size_t values_count);
 
 /** Get the index of a float array element which is closest to the given value.
  *
  * Returned index corresponds to the first element found.
- * If no suitable elements were found, the function returns 0.
+ * If no suitable elements were found, the function returns -1.
  *
  * @param   value           value to be searched.
  * @param   values          pointer to the array to perform the search in.
@@ -45,12 +45,12 @@ size_t value_index_uint32(const uint32_t value, const uint32_t values[], size_t 
  *
  * @return value's index.
  */
-size_t value_index_float(const float value, const float values[], size_t values_count);
+int32_t value_index_float(const float value, const float values[], size_t values_count);
 
 /** Get the index of a bool array element which is equal to the given value.
  *
  * Returned index corresponds to the first element found.
- * If no suitable elements were found, the function returns 0.
+ * If no suitable elements were found, the function returns -1.
  *
  * @param   value           value to be searched.
  * @param   values          pointer to the array to perform the search in.
@@ -58,12 +58,12 @@ size_t value_index_float(const float value, const float values[], size_t values_
  *
  * @return value's index.
  */
-size_t value_index_bool(const bool value, const bool values[], size_t values_count);
+int32_t value_index_bool(const bool value, const bool values[], size_t values_count);
 
 /** Get the index of a string array element which is equal to the given value.
  *
  * Returned index corresponds to the first element found.
- * If no suitable elements were found, the function returns 0.
+ * If no suitable elements were found, the function returns -1.
  *
  * @param   value           value to be searched.
  * @param   values          pointer to the array to perform the search in.
@@ -71,13 +71,13 @@ size_t value_index_bool(const bool value, const bool values[], size_t values_cou
  *
  * @return value's index.
  */
-size_t value_index_string(const char* value, const char* const values[], size_t values_count);
+int32_t value_index_string(const char* value, const char* const values[], size_t values_count);
 
 /**
  * @brief String-to-string mapping
  * 
  * Finds out the index of `input` in `src_array`, returns the element at the
- * same index in `dst_array`.
+ * same index in `dst_array`. Returns NULL if `input` is not found in `src_array`.
  * 
  * @param   src_array Array of source strings (to map from)
  * @param   dst_array Array of destination strings (to map to)
